requirements.txt
Deprecated==1.2.14
lxml==5.3.0
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.1
wrapt==1.16.0
